--- src/Altinn.Notifications.Core/AltinnServiceUpdate/AltinnService.cs
@@ -1,18 +1,0 @@
-﻿namespace Altinn.Notifications.Core.AltinnServiceUpdate
-{
-    /// <summary>
-    /// Enum representing the different Altinn services that can provide service updates
-    /// </summary>
-    public enum AltinnService
-    {
-        /// <summary>
-        /// The default service
-        /// </summary>
-        Unknown,
-
-        /// <summary>
-        /// The Altinn Notifications Email microservice
-        /// </summary>
-        Notifications_Email
-    }
-}
